move FieldName*Test to own file
