+++ corehq/apps/fixtures/tests/__init__.py
@@ -1,4 +1,4 @@
-from .tests import *
+from .test_field_names import *
 from .test_dbaccessors import *
 from .test_fixture_data import *
 from .test_location_ownership import *
